Fix incorrect computation of miss rates due to Int/Long mixups <sigh>.
git-svn-id: svn://svn.valgrind.org/valgrind/branches/ASPACEM@4768 a5019735-40e9-0310-863c-91ae7b9d1cf9

+++ include/pub_tool_libcprint.h
@@ -52,7 +52,8 @@ extern UInt VG_(vsnprintf)( Char* buf, Int size,
                                        const HChar *format, va_list vargs );
 
 // Percentify n/m with d decimal places.  Includes the '%' symbol at the end.
-extern void VG_(percentify)(UInt n, UInt m, UInt d, Int n_buf, char buf[]);
+// Right justifies in 'buf'.
+extern void VG_(percentify)(ULong n, ULong m, UInt d, Int n_buf, char buf[]);
 
 /* ---------------------------------------------------------------------
    Messages for the user
